+++ debug-adapter/app/Phoityne/VSCode/TH/LaunchRequestArgumentsJSON.hs
@@ -25,6 +25,9 @@ data LaunchRequestArguments =
   , stopOnEntryLaunchRequestArguments  :: Bool       -- Phoityne specific argument. required.
   , mainArgsLaunchRequestArguments     :: Maybe String         -- Phoityne specific argument. required.
   , ghciEnvLaunchRequestArguments      :: M.Map String String  -- Phoityne specific argument. required.
+  , ghciInitialPromptLaunchRequestArguments  :: Maybe String     -- Phoityne specific argument. required.
+  , startupFuncLaunchRequestArguments  :: Maybe String
+  , startupArgsLaunchRequestArguments  :: Maybe String
   } deriving (Show, Read, Eq)